+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 Every release, along with the migration instructions, is documented on the Github [Releases](https://github.com/FormidableLabs/webpack-dashboard/releases) page.
 
+## UNRELEASED
+
+- Add `DashboardPlugin({ includeAssets: [ "stringPrefix", /regexObj/ ] })` Webpack plugin filtering option.
+- Add `webpack-dashboard --include-assets stringPrefix1 -a stringPrefix1` CLI filtering option.
+- Change `"mode"` SocketIO event to `"options"` as it now passes both `minimal` and `includeAssets` from CLI to the Webpack plugin.
+
 ## [3.0.7] - 2019-05-15
 
 ### Features